Removed debug.lua Added function into DebugPlayer.lua
